BJS2-52983: avatar user

+++ src/main/java/school/faang/user_service/properties/AvatarProperties.java
@@ -19,8 +19,8 @@ public class AvatarProperties {
 
     private String contentType;
 
-    private long maxWidth;
+    private int maxWidth;
 
-    private long maxSmallWidth;
+    private int maxSmallWidth;
 
 }
